refactor(referrals): move referral domain primitives
